Instead of adding database CRUD, just directly manipulate the new UserSet column.
The database CRUD was inefficient and overkill.  This approach doesn't
need any additional database access when rendering a problem.  The user
set already has the necessary data.  Just use it.  When a problem is
submitted with this data the data is processed and added to the new
column.  This just requires obtaining a clean copy of the user set from
the database and then saving the data to the new column.  This is much
like the processing of answers for the problem.  Note that the new
course environment option for using external data is removed.  With this
approach the only time that additional database access is needed is if a
problem actually uses external data.

Note that the data for this is now not also stored in the last_answer
column for the problem.

The unit test is removed.  It just isn't needed.  That means the dev
conf file is also not needed.