Improve layouts and fix many issues.
Most text is now translatable.

The links no longer use the "html" extension.  The "html" extension is
only needed for actual html files.  Mojolicious does not need those.
This now uses properly resolved links for routes instead.

The problems with the clipboard icon are fixed.  That and several other
issues were caused by using css classes that are from bootstrap 5.3,
that are not in bootstrap 5.2.

This requires changes to the way the sample problems are parsed that are
in a pull request to the PG side of this.