Fixes and improvements for the new "Make Archive" page.
Fix the action.  There needs to be a hidden "confirm" input.

Improve layouts.

Also other clean up suggested in my review of #2163.

Don't use `chdir`.

The IO::Compression::Zip module is too limited.  It can not add
directories to the zip file.  Empty directories in particular should be.
So use Archive::Zip instead.  This module is already in use by webwork
as well.